add old_key support to bfile_attr.

+++ tests/lib/bes/files/metadata/_test_fruits_factory.py
@@ -13,5 +13,5 @@ class _test_fruits_factory(bfile_metadata_factory_base):
     return [
       ( 'acme/fruit/kiwi/1.0', lambda f: os.stat(f).st_size, clazz.decode_int, clazz.encode_int, check.check_int, None ),
       ( 'acme/fruit/cherry/2.0', lambda f: float(os.stat(f).st_size / 2.0), clazz.decode_float, clazz.encode_float, check.check_float, None ),
-      ( 'acme/fruit/melon/1.0', lambda f: float(os.stat(f).st_size * 2), clazz.decode_int, clazz.encode_int, check.check_int, ( 'bes_double_size', ) ),
+      ( 'acme/fruit/melon/1.0', lambda f: os.stat(f).st_size * 2, clazz.decode_int, clazz.encode_int, check.check_int, ( 'bes_double_size', ) ),
     ]
